fix several arm7l portability issues

+++ src/cmd/flux-ping.c
@@ -137,7 +137,7 @@ done:
     if (flux_rpc_next (rpc) < 0) {
         if (pdata->rpc_count) {
             if (ctx->rank_count > 1) {
-                printf ("%s!%s pad=%lu seq=%d time=(%0.3f:%0.3f:%0.3f) ms "
+                printf ("%s!%s pad=%zu seq=%d time=(%0.3f:%0.3f:%0.3f) ms "
                         "stddev %0.3f\n",
                         ctx->rank, ctx->topic, strlen (ctx->pad), pdata->seq,
                         tstat_min (tstat), tstat_mean (tstat),
@@ -176,8 +176,8 @@ void send_ping (struct ping_ctx *ctx)
     rpc = flux_rpcf_multi (ctx->h, ctx->topic, ctx->rank, 0,
                            "{s:i s:I s:I s:s}",
                            "seq", ctx->send_count,
-                           "time.tv_sec", t0.tv_sec,
-                           "time.tv_nsec", t0.tv_nsec,
+                           "time.tv_sec", (uint64_t)t0.tv_sec,
+                           "time.tv_nsec", (uint64_t)t0.tv_nsec,
                            "pad", ctx->pad);
     if (!rpc)
         log_err_exit ("flux_rpcf_multi");
